Move disableMP argument from transformExecutor class to athenaExecutor class (ATLASJT-318) (PyJobTransforms-03-00-09)

+++ Tools/PyJobTransforms/python/trfMPTools.py
@@ -4,7 +4,7 @@
 #
 # @brief Utilities for handling AthenaMP jobs
 # @author atlas-comp-transforms-dev@cern.ch
-# @version $Id: trfMPTools.py 677748 2015-06-23 20:29:35Z graemes $
+# @version $Id: trfMPTools.py 750283 2016-05-27 12:30:08Z graemes $
 # 
 
 __version__ = '$Revision'
@@ -36,16 +36,17 @@ def detectAthenaMPProcs(argdict = {}):
                 raise ValueError("ATHENA_PROC_NUMBER value was less than zero")
             msg.info('AthenaMP detected from ATHENA_PROC_NUMBER with {0} workers'.format(athenaMPProcs))
         elif 'athenaopts' in argdict:
-            procArg = [opt.replace("--nprocs=", "") for opt in argdict['athenaopts'].value if '--nprocs' in opt]
-            if len(procArg) == 0:
-                athenaMPProcs = 0
-            elif len(procArg) == 1:
-                athenaMPProcs = int(procArg[0])
-                if athenaMPProcs < 0:
-                    raise ValueError("--nprocs was set to a value less than zero")
-            else:
-                raise ValueError("--nprocs was set more than once in 'athenaopts'")
-            msg.info('AthenaMP detected from "nprocs" setting with {0} workers'.format(athenaMPProcs))
+            for substep in argdict['athenaopts'].value:
+                procArg = [opt.replace("--nprocs=", "") for opt in argdict['athenaopts'].value[substep] if '--nprocs' in opt]
+                if len(procArg) == 0:
+                    athenaMPProcs = 0
+                elif len(procArg) == 1:
+                    athenaMPProcs = int(procArg[0])
+                    if athenaMPProcs < 0:
+                        raise ValueError("--nprocs was set to a value less than zero")
+                else:
+                    raise ValueError("--nprocs was set more than once in 'athenaopts'")
+                msg.info('AthenaMP detected from "nprocs" setting with {0} workers for substep {1}'.format(athenaMPProcs,substep))
     except ValueError, errMsg:
         myError = 'Problem discovering AthenaMP setup: {0}'.format(errMsg)
         raise trfExceptions.TransformExecutionException(trfExit.nameToCode('TRF_EXEC_SETUP_FAIL'), myError)
@@ -72,14 +73,14 @@ def athenaMPOutputHandler(athenaMPFileReport, athenaMPWorkerTopDir, dataDictiona
     for filesElement in mpOutputs.getroot().getiterator(tag='Files'):
         msg.debug('Examining element {0} with attributes {1}'.format(filesElement, filesElement.attrib))
         originalArg = None 
-        originalName = filesElement.attrib['OriginalName']
+        startName = filesElement.attrib['OriginalName']
         for dataType, fileArg in dataDictionary.iteritems():
-            if fileArg.value[0] == originalName:
+            if fileArg.value[0] == startName:
                 originalArg = fileArg
                 outputHasBeenHandled[dataType] = True
                 break
         if originalArg is None:
-            msg.warning('Found AthenaMP output with name {0}, but no matching transform argument'.format(originalName))
+            msg.warning('Found AthenaMP output with name {0}, but no matching transform argument'.format(startName))
             continue
         
         msg.debug('Found matching argument {0}'.format(originalArg))
@@ -101,14 +102,14 @@ def athenaMPOutputHandler(athenaMPFileReport, athenaMPWorkerTopDir, dataDictiona
             if fileArg.io is "input":
                 continue
             msg.info("Searching MP worker directories for {0}".format(dataType))
-            originalName = fileArg.value[0]
+            startName = fileArg.value[0]
             fileNameList = []
             for entry in MPdirWalk:
                 if "evt_count" in entry[0]:
                     continue
                 # N.B. AthenaMP may have made the output name unique for us, so 
                 # we need to treat the original name as a prefix
-                possibleOutputs = [ fname for fname in entry[2] if fname.startswith(originalName) ]
+                possibleOutputs = [ fname for fname in entry[2] if fname.startswith(startName) ]
                 if len(possibleOutputs) == 0:
                     continue
                 elif len(possibleOutputs) == 1:
@@ -123,33 +124,32 @@ def athenaMPOutputHandler(athenaMPFileReport, athenaMPWorkerTopDir, dataDictiona
 
 
 def athenaMPoutputsLinkAndUpdate(newFullFilenames, fileArg):
-    originalName = fileArg.value[0]
-    # Do we need to append worker dir suffixes?
+    # Any files we link are numbered from 1, because we always set
+    # the filename given to athena has _000 as a suffix so that the
+    # mother process' file can be used without linking
+    fileIndex = 1
     linkedNameList = []
-    uniqueSimpleNames = set([path.basename(fname) for fname in newFullFilenames])
-    if len(uniqueSimpleNames) != len(newFullFilenames):
-        for fname in newFullFilenames:
-            simpleName = path.basename(fname)
-            workerIndexMatch = re.search(r'/worker_(\d+)/', fname)
-            if workerIndexMatch:
-                workerIndex = workerIndexMatch.group(1)
-            else:
-                raise trfExceptions.TransformExecutionException(trfExit.nameToCode("TRF_OUTPUT_FILE_ERROR"), "Found output file ({0}) not in an AthenaMP worker directory".format(fname))
-            simpleName += "._{0:03d}".format(int(workerIndex))
-            linkedNameList.append(simpleName)
-    else:
-        linkedNameList = [path.basename(fname) for fname in newFullFilenames]
+    newFilenameValue = []
+    for fname in newFullFilenames:
+        if path.dirname(fname) == "":
+            linkedNameList.append(None)
+            newFilenameValue.append(fname)
+        else:
+            linkName = "{0}{1:03d}".format(path.basename(fname).rstrip('0'), fileIndex)
+            linkedNameList.append(linkName)
+            newFilenameValue.append(linkName)
+            fileIndex += 1
 
     for linkname, fname in zip(linkedNameList, newFullFilenames):
-        try:
-            if path.lexists(linkname):
-                os.unlink(linkname)
-            os.symlink(fname, linkname)
-        except OSError, e:  
-            raise trfExceptions.TransformExecutionException(trfExit.nameToCode("TRF_OUTPUT_FILE_ERROR"), "Failed to link {0} to {1}: {2}".format(fname, linkname, e))
+        if linkname:
+            try:
+                if path.lexists(linkname):
+                    os.unlink(linkname)
+                os.symlink(fname, linkname)
+            except OSError, e:  
+                raise trfExceptions.TransformExecutionException(trfExit.nameToCode("TRF_OUTPUT_FILE_ERROR"), "Failed to link {0} to {1}: {2}".format(fname, linkname, e))
 
     fileArg.multipleOK = True
-    fileArg.value = linkedNameList
-    fileArg.originalName = originalName
+    fileArg.value = newFilenameValue
     msg.debug('MP output argument updated to {0}'.format(fileArg))
